# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface EnterpriseRisk {
  eal: number;
  var95: number;
  cvar95: number;
  likelihood: number;
  riskScore: number;
  delta: number;
  eventCount: number;
}

export interface Overview {
  organization: string;
  enterprise: EnterpriseRisk;
  criticalAssets: number;
  controlEffectiveness: number;
  lastUpdated: string;
  eventCount: number;
}

export interface BusinessUnitRisk {
  id: string;
  name: string;
  eal: number;
  riskScore: number;
  assetCount: number;
  criticalAssets: number;
  delta: number;
}

export interface ServiceRisk {
  id: string;
  name: string;
  businessUnit: string;
  eal: number;
  riskScore: number;
  criticality: number;
}

export interface AssetRisk {
  id: string;
  name: string;
  service: string;
  businessUnit: string;
  assetType: string;
  criticality: number;
  eal: number;
  likelihood: number;
  controlEffectiveness: number;
  severity: string;
  internetExposed: boolean;
}

export interface RiskDriver {
  id: string;
  name: string;
  category: string;
  contribution: number;
  share: number;
  severity: string;
  affectedAssets: number;
  explanation: string;
}

export interface RiskTrend {
  label: string;
  eal: number;
  riskScore: number;
  eventCount: number;
}

export type Asset = AssetRisk;

export interface Vulnerability {
  id: string;
  cve: string;
  title: string;
  cvss: number;
  severity: string;
  exploitAvailable: boolean;
  activelyExploited: boolean;
  affectedAssets: number;
  ealContribution: number;
  remediationStatus: string;
  ageDays: number;
}

export interface Control {
  id: string;
  name: string;
  category: string;
  coverage: number;
  configuration: number;
  compliance: number;
  telemetry: number;
  effectiveness: number;
  status: string;
}

export interface Contribution {
  label: string;
  value: number;
  direction: string;
}

export interface MitigationAction {
  id: string;
  name: string;
  category: string;
  cost: number;
  reduction: number;
  days: number;
  description: string;
}

export type AssetDetail = AssetRisk & {
  owner: string;
  environment: string;
  dataSensitivity: string;
  vulnerabilities: Vulnerability[];
  controls: Control[];
  dependencies: string[];
  contributions: Contribution[];
  recommendations: MitigationAction[];
};

export type VulnerabilityDetail = Vulnerability & {
  affectedServices: string[];
  affectedBusinessUnits: string[];
  controls: Control[];
  recommendation: string;
};

export interface Incident {
  id: string;
  type: string;
  severity: string;
  date: string;
  asset: string;
  totalLoss: number;
  status: string;
}

export interface Threat {
  id: string;
  name: string;
  category: string;
  activityLevel: number;
  exploitability: number;
  targetSector: string;
}

export interface ScenarioInput {
  actionIds: string[];
  mfaCoverage?: number;
  delayDays?: number;
}

export interface CurvePoint {
  label: string;
  value: number;
}

export interface ScenarioResult {
  baselineEal: number;
  projectedEal: number;
  reduction: number;
  cost: number;
  rosi: number;
  actionIds: string[];
  exposureCurve: CurvePoint[];
}

export interface OptimizationInput {
  /** @minimum 0 */
  budget: number;
}

export interface OptimizationResult {
  budget: number;
  totalCost: number;
  reduction: number;
  remainingBudget: number;
  actions: MitigationAction[];
  curve: CurvePoint[];
}

export interface ComplianceRequirement {
  code: string;
  title: string;
  control: string;
  status: string;
  evidence: string;
  lastVerified: string;
  gap: string;
  remediation: string;
}

export interface ComplianceFramework {
  id: string;
  name: string;
  version: string;
  posture: number;
  requirements: ComplianceRequirement[];
  gaps: number;
}

export interface IngestEventInput {
  type: string;
  assetId?: string;
  severity?: string;
}

export interface IngestResult {
  message: string;
  eventType: string;
  ealBefore: number;
  ealAfter: number;
  delta: number;
  driver: RiskDriver;
  timeline: string[];
}

export interface CopilotInput {
  /** @minLength 1 */
  question: string;
}

export interface CopilotResponse {
  answer: string;
  citations: string[];
  tool: string;
}

export interface Notification {
  id: string;
  type: string;
  title: string;
  detail: string;
  time: string;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type BusinessUnitParameter = string;

export type SeverityParameter = string;

export type GetAssetRiskListParams = {
businessUnit?: BusinessUnitParameter;
severity?: SeverityParameter;
};