# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetRisk } from './assetRisk';

export type Asset = AssetRisk;