# lib/api-zod/src/generated/types/assetDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetRisk } from './assetRisk';
import type { Contribution } from './contribution';
import type { Control } from './control';
import type { MitigationAction } from './mitigationAction';
import type { Vulnerability } from './vulnerability';

export type AssetDetail = AssetRisk & {
  owner: string;
  environment: string;
  dataSensitivity: string;
  vulnerabilities: Vulnerability[];
  controls: Control[];
  dependencies: string[];
  contributions: Contribution[];
  recommendations: MitigationAction[];
};