# lib/api-zod/src/generated/types/assetRisk.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface AssetRisk {
  id: string;
  name: string;
  service: string;
  businessUnit: string;
  assetType: string;
  criticality: number;
  eal: number;
  likelihood: number;
  controlEffectiveness: number;
  severity: string;
  internetExposed: boolean;
}