# lib/api-zod/src/generated/types/businessUnitRisk.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessUnitRisk {
  id: string;
  name: string;
  eal: number;
  riskScore: number;
  assetCount: number;
  criticalAssets: number;
  delta: number;
}