# lib/api-zod/src/generated/types/control.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface Control {
  id: string;
  name: string;
  category: string;
  coverage: number;
  configuration: number;
  compliance: number;
  telemetry: number;
  effectiveness: number;
  status: string;
}