# lib/api-zod/src/generated/types/copilotInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface CopilotInput {
  /** @minLength 1 */
  question: string;
}