# lib/api-zod/src/generated/types/copilotResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface CopilotResponse {
  answer: string;
  citations: string[];
  tool: string;
}