# lib/api-zod/src/generated/types/ingestEventInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface IngestEventInput {
  type: string;
  assetId?: string;
  severity?: string;
}