# lib/api-zod/src/generated/types/optimizationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
import type { CurvePoint } from './curvePoint';
import type { MitigationAction } from './mitigationAction';

export interface OptimizationResult {
  budget: number;
  totalCost: number;
  reduction: number;
  remainingBudget: number;
  actions: MitigationAction[];
  curve: CurvePoint[];
}