# lib/api-zod/src/generated/types/riskTrend.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskTrend {
  label: string;
  eal: number;
  riskScore: number;
  eventCount: number;
}