# lib/api-zod/src/generated/types/scenarioResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
import type { CurvePoint } from './curvePoint';

export interface ScenarioResult {
  baselineEal: number;
  projectedEal: number;
  reduction: number;
  cost: number;
  rosi: number;
  actionIds: string[];
  exposureCurve: CurvePoint[];
}