# lib/api-zod/src/generated/types/serviceRisk.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceRisk {
  id: string;
  name: string;
  businessUnit: string;
  eal: number;
  riskScore: number;
  criticality: number;
}