# lib/api-zod/src/generated/types/threat.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Threat Quotient synthetic cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */

export interface Threat {
  id: string;
  name: string;
  category: string;
  activityLevel: number;
  exploitability: number;
  targetSector: string;
}